+++ 3dprinter/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+[2021.11.2]
+
+- Add validation for OctoPrint command, with automatic attempt at repair on fail.
+
+[2021.11.1]
+
+- Add repair option
+- Fix Klipper update script
+
 [2021.11.0]
 
 - Use custom `moonraker.conf` if available